+++ plugin.go
@@ -226,7 +226,7 @@ func mergeOverride[F any](
 	getter func(testoplugin.Overrides) testoplugin.Override[F],
 ) func(F) F {
 	return func(f F) F {
-		for _, o := range overrides {
+		for _, o := range slices.Backward(overrides) {
 			if override := getter(o); override != nil {
 				f = override(f)
 			}